Atualização do service

+++ backend/src/todos/use-cases/delete-todo.use-case.ts
@@ -1,5 +1,6 @@
-import { Injectable, Logger } from "@nestjs/common";
+import { Injectable, Logger, NotFoundException } from "@nestjs/common";
 import { DeleteTodoRepository } from "../repositoy";
+import { FindTodoByIDRepository } from "../repositoy";
 
 
 
@@ -7,15 +8,25 @@ import { DeleteTodoRepository } from "../repositoy";
 export class DeleteTodoUseCase{
     constructor(
         private readonly deleteTodoRepository: DeleteTodoRepository,
+        private readonly findtodobyidRepository: FindTodoByIDRepository,
         private readonly logger: Logger, 
     ) {}
 
     async execute(id: string){
         try{
             this.logger.log('Deleting toDo...');
-            const todo = await this.deleteTodoRepository.delete(id);
+
+            const todos = await this.findtodobyidRepository.FindById(id)
+
+            if (!todos) {
+                throw new NotFoundException('ToDo not found');
+            }
+
+            await this.deleteTodoRepository.delete(id);
+
             this.logger.log('ToDo Deleted sucessfully');
-            return todo;
+            
+            return todos;
         }
         catch(error) {
             this.logger.error(error);
